chore(suite): merge TransactionReviewModalContent with TransactionReviewModal
